+++ README.md
@@ -37,3 +37,15 @@
 ## 부록
 ### 생각 로그
 - 요구사항 정의 목표: 유저가 게임 동작 과정을 이해할 수 있게 작성한다.
+
+### 고민한 포인트
+- 의존성 갖는 관계에서 매개변수로 객체를 넘기는 것과, 필드로 객체를 갖는 것의 차이점
+  - 매개변수
+    - 장점: 객체를 메시지처럼 주고받을 수 있음 (사용하는 측에서 단일 객체를 쓰던, 객체 리스트를 쓰던 상관 없이 사용 가능함)
+      - 상관 없이 사용 가능하단 게 잘 안와닿긴 함 (ex. car를 사용하는 racing이 있다고 할 때)
+        - racing이 car를 매개변수로 받을 땐 List<Car>를 가진 누군가가 racing을 사용할 때 racing 객체를 하나만 만들고, 함수를 여러 번 호출하면 됨
+        - racing이 car를 필드로 받을 땐 List<Car>를 가진 누군가가 racing을 사용할 때 List<Racing> 처럼 객체를 여러 개 만들고, 각각의 함수를 한 번씩 호출하면 됨
+    - 단점: 동일한 매개변수를 메서드마다 반복해서 넣어주는 일 생기기도 함 (매개변수가 많아짐)
+  - 필드
+    - 장점: 모든 메서드가 사용하는 필드의 경우, 중복 코드 줄일 수 있음
+    - 단점: 필드 형태가 고정되어 있어 유연하지 못함
