+++ src/main/java/model/Cars.java
@@ -3,6 +3,7 @@ package model;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
+import model.dto.CarState;
 
 public class Cars {
     private final List<Car> cars;
