+++ src/main/java/model/dto/CarState.java
@@ -1,4 +1,6 @@
-package model;
+package model.dto;
+
+import model.Car;
 
 public record CarState(String name, int forwardCount) {
     public static CarState create(Car car) {
